Growbinar_Backend/Authentication/serializers.py
from rest_framework import serializers
from static.models import Mentee,Mentor

class MentorSerializer(serializers.ModelSerializer):
    class Meta:
        model = Mentor
        fields = '__all__'



class MenteeSerializer(serializers.ModelSerializer):
    class Meta:
        model = Mentee
        fields = '__all__'